requirements.txt
python-telegram-bot[ext]
python-dotenv
torch
transformers
Pillow
